modified:   backend/api_directory/schemas.py 	modified:   backend/database_directory/models.py 	modified:   frontend/app.py 	modified:   frontend/static/dashboard.css 	modified:   frontend/telemetry/export/pdf.py 	deleted:    frontend/telemetry/export/templates/average.html 	deleted:    frontend/telemetry/export/templates/overview.html 	deleted:    frontend/telemetry/export/templates/session.html 	modified:   frontend/telemetry/export/templates/styles.css 	new file:   frontend/telemetry/export/templates/template.html 	modified:   frontend/telemetry/initialize/file_analyser.py 	modified:   frontend/telemetry/modules/sorting.py 	modified:   frontend/templates/dashboard.html

+++ backend/api_directory/schemas.py
@@ -42,7 +42,7 @@ class BoatTelemetryData(BaseModel):
     normalizedcatch: List[float]
     normalizedfinish: List[float]
     sample_time: List[float] 
-    sample_strokes: List[float]
+    sample_strokes: List[int]
     
     # 2D arrays
     acceleration: List[List[float]]
